Fix missing method controls, hopefully save a bit of memory
The memory saving is changing AddObjectControls to not use bool arrays to determine which controls are methods and which are updates. I changed it to being a ulong and just checking bit indexes. It makes the code a bit more confusing to look at but hopefully way more memory efficient.

FVRGrenade has been deprecated as well so I just switched it to PinnedGrenade.